requirements.txt
astroid==2.5
click==7.1.2
colorama==0.4.4
Flask==1.1.2
gunicorn==20.0.4
isort==5.7.0
itsdangerous==1.1.0
Jinja2==2.11.3
lazy-object-proxy==1.5.2
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.18.4
opencv-python==4.2.0.34
Pillow==8.1.0
pybullet==2.7.5
pylint==2.7.1
pytesseract==0.3.7
toml==0.10.2
typed-ast==1.4.2
Werkzeug==1.0.1
wrapt==1.12.1
